local/binary/OppoDecrypt-main/pyproject.toml
[tool.poetry]
name = "oppodecrypt"
version = "0.1.0"
description = "Oppo .ofp and Oneplus .ops Firmware decrypter"
authors = ["Sergey K <unix3dgforce@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
dependency-injector = "^4.41.0"
loguru = "^0.7.0"
pyyaml = "^6.0"
pycryptodome = "^3.18.0"
pycryptodomex = "^3.18.0"
inquirerpy = "^0.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
